Fix alert infinite loop on repeat interval of 0 (#52628)
* #4851 - Infinite loop on repeat interval of 0

Notification will enter an infinite loop when the repeat interval is specified as zero and it is the last repeat configured.  When this occurs avoid the infinite loop and log a warning message.  Note: I encountered this issue when routing SMS to Twilio and quickly sent thousands of text messages.

* Update __init__.py

* Remove runtime check since configuration input is now blocked

* Tweak comment

Co-authored-by: Franck Nijhof <frenck@frenck.nl>